code cleanup moved components to ds-prototypes
