+++ src/styles/main.scss
@@ -15,7 +15,7 @@
 @import './../components/location-finder/Hero/Hero.scss';
 @import './../components/location-finder/SearchForm/SearchForm.scss';
 @import './../components/location-finder/SearchFilters/SearchFilters.scss';
-@import './../components/location-finder/SearchAutoSuggest/SearchAutoSuggest.scss';
+@import './../components/ds-prototypes/AutoSuggest/AutoSuggest.scss';
 @import './../components/location-finder/SearchHeader/SearchHeader.scss';
 @import './../components/location-finder/Locations/Locations.scss';
 @import './../components/location-finder/Location/Location.scss';
@@ -23,7 +23,7 @@
 @import './../components/location-finder/BottomPromo/BottomPromo.scss';
 @import './../components/location-finder/RightRail/RightRail.scss';
 @import './../components/shared/Dropdown/Dropdown.scss';
-@import './../components/shared/CardGrid/CardGrid.scss';
+@import './../components/ds-prototypes/CardGrid/CardGrid.scss';
 
 
 // shared/layouts > Main.js
